Reworking the 3d room
## Different approach to making a 3d scalable room.
It seems the current script scales not only the room itself, but also the thickness of the walls. Warping the entire room. On top of that, currently setting the room to a certain scale using the metric system scales the room times it's original size, resulting in what's suppoused to be a 8x8 m room, being  48x48 meters. and the higher the number the more ridiculous it gets.
Therefore we are reworking the entire idea of how the room is made. by making a plane, and anchoring walls to the middle edge of each side of the plane. So width is gonna control the plane and 2 walls, length is gonna control the plane, and 2 walls. And height is only gonna control the walls

Co-Authored-By: huniche <91188216+huniche@users.noreply.github.com>